backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
openai==2.8.1
python-dotenv==1.0.0

